Move useColorTheme
- Move to hooks folder

- Update deps import statements
