+++ src/components/app.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { ColorThemeContextProvider } from '@/styles/useColorTheme';
+import { ColorThemeContextProvider } from '@/hooks/useColorTheme';
 
 /**
  * Wraps all pages on the site.
